Create read and delete selection commands
