+++ addon/commands/read-selection-command.ts
@@ -1,0 +1,6 @@
+import SelectionCommand from "@lblod/ember-rdfa-editor/commands/selection-command";
+
+export default class ReadSelectionCommand extends SelectionCommand {
+  name = "read-selection";
+  deleteSelection = false;
+}
