+++ addon/model/model.ts
@@ -11,18 +11,16 @@ import BatchedModelMutator from "@lblod/ember-rdfa-editor/model/mutators/batched
 import ImmediateModelMutator from "@lblod/ember-rdfa-editor/model/mutators/immediate-model-mutator";
 import ModelRange from "@lblod/ember-rdfa-editor/model/model-range";
 
-
 /**
  * Abstraction layer for the DOM. This is the only class that is allowed to call DOM methods.
  * Code that needs to modify the DOM has to use a {@link Command}.
  * The model is still exposed for querying but that might become even more restricted later.
  */
 export default class Model {
-
   /**
    * The root of the editor. This will get set by ember,
-   * so we trick typescript into assuming it is never null
-   * @private
+   * so we trick typescript into assuming it is never null.
+   * @protected
    */
   protected _rootModelNode!: ModelElement;
   private reader: HtmlReader;
@@ -47,7 +45,6 @@ export default class Model {
     return this._rootNode;
   }
 
-
   get selection(): ModelSelection {
     return this._selection;
   }
@@ -56,9 +53,8 @@ export default class Model {
     return this._rootModelNode;
   }
 
-
   /**
-   * Read in the document and build up the model
+   * Read in the document and build up the model.
    */
   read(readSelection = true) {
     const parsedNodes = this.reader.read(this.rootNode);
@@ -67,11 +63,12 @@ export default class Model {
     }
     const newRoot = parsedNodes[0];
     if (!ModelNode.isModelElement(newRoot)) {
-      throw new Error("root model node has to be an element");
+      throw new Error("Root model node has to be an element");
     }
     this._rootModelNode = newRoot;
     this.bindNode(this.rootModelNode, this.rootNode);
-    // This is essential, we change the root so we need to make sure the selection uses the new root
+
+    // This is essential, we change the root so we need to make sure the selection uses the new root.
     if (readSelection) {
       this.readSelection();
     }
@@ -81,31 +78,34 @@ export default class Model {
     this._selection = this.selectionReader.read(domSelection);
   }
 
-  /**]
-   * Write a part of the model back to the dom
+  /**
+   * Write a part of the model back to the dom.
    * @param tree
-   * @param writeSelection if we should also write out the selection. Should be
+   * @param writeSelection If we should also write out the selection. Should be
    * almost always true, but can be useful for testing to turn it off when you dont
-   * have a real dom available
+   * have a real dom available.
    */
   write(tree: ModelElement = this.rootModelNode, writeSelection = true) {
-    const modelWriteEvent = new CustomEvent(
-      'editorModelWrite',
-    );
+    const modelWriteEvent = new CustomEvent("editorModelWrite");
     document.dispatchEvent(modelWriteEvent);
+
     const oldRoot = tree.boundNode;
     if (!oldRoot) {
-      throw new Error("Container without boundNOde");
+      throw new Error("Container without boundNode");
     }
+
     if (!isElement(oldRoot)) {
       throw new NotImplementedError("root is not an element, not sure what to do");
     }
+
     const newRoot = this.writer.write(tree);
     while (oldRoot.firstChild) {
       oldRoot.removeChild(oldRoot.firstChild);
     }
+
     oldRoot.append(...newRoot.childNodes);
     this.bindNode(tree, oldRoot);
+
     if (writeSelection) {
       this.writeSelection();
     }
@@ -115,10 +115,9 @@ export default class Model {
     this.selectionWriter.write(this.selection);
   }
 
-
   /**
    * Bind a modelNode to a domNode. This ensures that we can reach the corresponding node from
-   * either side
+   * either side.
    * @param modelNode
    * @param domNode
    */
@@ -129,46 +128,56 @@ export default class Model {
   }
 
   /**
-   * Get the corresponding modelNode for domNode
+   * Get the corresponding modelNode for domNode.
    * @param domNode
    */
   public getModelNodeFor(domNode: Node): ModelNode {
     if (!this.nodeMap) throw new ModelError("uninitialized nodeMap");
-    const rslt = this.nodeMap.get(domNode);
-    if (!rslt) {
+
+    const result = this.nodeMap.get(domNode);
+    if (!result) {
       throw new ModelError("No boundnode for domNode");
     }
-    return rslt;
+
+    return result;
   }
 
   /**
-   * Change the model by providing a callback with will receive an {@link ImmediateModelMutator immediate mutator}
+   * Change the model by providing a callback that will receive an {@link ImmediateModelMutator immediate mutator}.
    * The model gets written out automatically after the callback finishes.
+   *
    * @param callback
+   * @param writeBack
    */
-  change(callback: (mutator: ImmediateModelMutator) => ModelElement | void) {
+  change(callback: (mutator: ImmediateModelMutator) => ModelElement | void, writeBack = false) {
     const mutator = new ImmediateModelMutator();
     const subTree = callback(mutator);
-    if (subTree) {
-      this.write(subTree);
-    } else {
-      this.write(this.rootModelNode);
+
+    if (writeBack) {
+      if (subTree) {
+        this.write(subTree);
+      } else {
+        this.write(this.rootModelNode);
+      }
     }
   }
 
-
   /**
-   * Change the model by providing a callback with will receive a {@link BatchedModelMutator batched mutator}
+   * Change the model by providing a callback that will receive a {@link BatchedModelMutator batched mutator}
    * The mutator gets flushed and the model gets written out automatically after the callback finishes.
+   *
    * @param callback
+   * @param autoSelect
    */
   batchChange(callback: (mutator: BatchedModelMutator) => ModelElement | void, autoSelect = true) {
     const mutator = new BatchedModelMutator();
     const subTree = callback(mutator);
+
     const resultingRange = mutator.flush();
     if (autoSelect && resultingRange) {
       this.selection.selectRange(resultingRange);
     }
+
     if (subTree) {
       this.write(subTree);
     } else {
@@ -181,14 +190,16 @@ export default class Model {
     if (!parent) {
       return null;
     }
+
+    // More verbose but probably more efficient than converting to an array and using indexOf.
     let index = 0;
-    // more verbose but probably more efficient than converting to an array and using indexOf
     for (const candidate of parent.childNodes) {
       if (child === candidate) {
         return index;
       }
       index++;
     }
+
     return null;
   }
 
@@ -199,5 +210,4 @@ export default class Model {
   toXml(): Node {
     return this.rootModelNode.toXml();
   }
-
 }
